Fix test-pattern mismatch for streams with more than one string channel

+++ src/sample.cpp
@@ -309,7 +309,7 @@ sample &sample::assign_test_pattern(int offset) {
 		break;
 	case cft_string: {
 		std::string *data = (std::string *)&data_;
-		for (uint32_t k = 0u; k < num_channels_; k++)
+		for (int32_t k = 0u; k < (int) num_channels_; k++)
 			data[k] = to_string((k + 10) * (k % 2 == 0 ? 1 : -1));
 		break;
 	}
